+++ TA Signals/SelectionView.swift
@@ -7,12 +7,18 @@
 
 import SwiftUI
 
-struct TempEmptyView: View {
+struct SelectionView: View {
+    
+    @StateObject var appSettings = TASettings()
+    
     var body: some View {
+        // TODO: repeated code, I can move this to AppConstants or TASettings:
+        let marketStatus = appSettings.isMarketOpen ? "Open 🟢" : "Closed 🔴"
+        
         TabView{
             NavigationView{
                 VStack{
-                    Text("Hello")
+                    Text("Hello. Market is \(marketStatus)")
                 }
                 .navigationTitle("TA Signals - Empty View")
                 .navigationBarTitleDisplayMode(.inline)
@@ -23,6 +29,6 @@ struct TempEmptyView: View {
 
 struct TempEmptyView_Previews: PreviewProvider {
     static var previews: some View {
-        TempEmptyView()
+        SelectionView()
     }
 }
